--- mvseg3d/utils/io_utils.py
@@ -1,14 +1,0 @@
-import torch
-import numpy as np
-
-
-def load_data_to_gpu(data_dict):
-    for key, val in data_dict.items():
-        if not isinstance(val, np.ndarray):
-            continue
-        else:
-            if key in ['point_indices', 'point_voxel_ids', 'labels']:
-                data_dict[key] = torch.from_numpy(val).long().cuda()
-            else:
-                data_dict[key] = torch.from_numpy(val).float().cuda()
-    return data_dict
