+++ tools/eval.py
@@ -10,7 +10,7 @@ from mvseg3d.datasets import build_dataloader
 from mvseg3d.models.segmentors.spnet import SPNet
 from mvseg3d.core.metrics import IOUMetric
 from mvseg3d.utils.logging import get_logger
-from mvseg3d.utils.io_utils import load_data_to_gpu
+from mvseg3d.utils.data_utils import load_data_to_gpu
 
 
 def parse_args():
